+++ frontend/src/components/ChatList.js
@@ -7,7 +7,7 @@ import ListItemAvatar from '@material-ui/core/ListItemAvatar';
 import Avatar from '@material-ui/core/Avatar';
 import ListItemText from '@material-ui/core/ListItemText';
 
-const DisplayDataWithList = ({ listData, icon, click, ...props }) => 
+const ChatList = ({ listData, icon, click, ...props }) => 
   <List
     hidden={props.value !== props.index}
     style={{ padding: 0 }}
@@ -36,4 +36,4 @@ const DisplayDataWithList = ({ listData, icon, click, ...props }) =>
     )}
   </List>
 
-export default DisplayDataWithList;
+export default ChatList;
