+++ frontend/src/components/MessagesContainer.js
@@ -15,7 +15,7 @@ import AccountCircleIcon from '@material-ui/icons/AccountCircle';
 import SendRoundedIcon from '@material-ui/icons/SendRounded';
 import LinearProgress from '@material-ui/core/LinearProgress';
 import Snackbar from '@material-ui/core/Snackbar';
-import UserDrawer from './UserDrawer';
+import ChatMemberDrawer from './ChatMemberDrawer';
 import { ChatInfoContext } from '../context/chatInfo';
 import AppBarWithTitle from './AppBarWithTitle';
 import { ReactComponent as HomeSvg } from '../images/home.svg';
@@ -130,7 +130,7 @@ function MessageList({ list, submit, socket })  {
     <Grid container direction="column" className={classes.root}>
       <Grid item className={classes.messageBarContainer}>
         <AppBarWithTitle title={chatInfo.chatName}>
-          <UserDrawer />
+          <ChatMemberDrawer />
         </AppBarWithTitle>
       </Grid>
       
@@ -169,7 +169,7 @@ function MessageList({ list, submit, socket })  {
         }
       </Grid>
       <Snackbar open={typing.start} message={`${typing.name} is typing`} style={{ bottom: '20%' }}/>
-      <Grid item className={classes.sendBarContainer}>
+      <Grid item className={classes.sendBarContainer} hidden={chatInfo.chatId ? false : true}>
         <Paper 
           component="form" 
           className={classes.sendBar}
